requirements.txt
asgiref==3.5.2
autodesk-forge-sdk==0.1.2
certifi==2022.6.15
charset-normalizer==2.1.0
dj-database-url==1.0.0
Django==4.1
django-environ==0.9.0
django-heroku==0.3.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
python-dotenv==0.20.0
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
whitenoise==6.2.0
